Tweak: full Compile() should make a smart decision about whether to use a `Type[]`, `Dictionary<Type,int>` or `Hashtable` for implementing `GetKeyImpl`
git-svn-id: http://protobuf-net.googlecode.com/svn/trunk@607 b72047a4-3652-0410-9617-b3994939e97b

+++ protobuf-net.unittest/PEVerify.cs
@@ -12,9 +12,13 @@ namespace ProtoBuf.unittest
     {
         public static void Verify(string path)
         {
-            Verify(path, 0);
+            Verify(path, 0, true);
         }
         public static void Verify(string path, int exitCode)
+        {
+            Verify(path, 0, true);
+        }
+        public static void Verify(string path, int exitCode, bool deleteOnSuccess)
         {
             // note; PEVerify can be found %ProgramFiles%\Microsoft SDKs\Windows\v6.0A\bin
             const string exePath = "PEVerify.exe";
@@ -26,7 +30,7 @@ namespace ProtoBuf.unittest
                 if (proc.WaitForExit(10000))
                 {
                     Assert.AreEqual(exitCode, proc.ExitCode, path);
-                    File.Delete(path);
+                    if(deleteOnSuccess) File.Delete(path);
                 }
                 else
                 {
